simplify the TripPlanner View logic into one Extension View
